fix: move postprocess to a single module, update ts tutorial
