+++ pnpxai/evaluator/metrics/pixel_flipping.py
@@ -8,7 +8,7 @@ from pnpxai.explainers.types import Tensor, TensorOrTupleOfTensors
 from pnpxai.explainers.base import Explainer
 from pnpxai.explainers import GradCam
 from pnpxai.utils import format_into_tuple, format_into_tuple_all
-from pnpxai.explainers.utils.postprocess.postprocess import PostProcessor
+from pnpxai.explainers.utils.postprocess import PostProcessor
 from pnpxai.evaluator.metrics.base import Metric
